requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1

# Task Queue
celery==5.3.4
python-celery-beat==2.5.0

# LLM & AI
langchain>=0.1.0
langchain-community>=0.0.10
langchain-ollama>=0.1.0
crewai>=0.28.0
ollama>=0.1.0
transformers>=4.35.2
torch==2.1.1
sentence-transformers==2.2.2

# Browser Automation
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Computer Vision & OCR
opencv-python==4.8.1.78
pytesseract==0.3.10
pillow==10.1.0
numpy==1.26.2

# Utilities
requests==2.31.0
pydantic==2.5.1
pydantic-settings==2.1.0
pyjwt==2.8.1
bcrypt==4.1.1
python-dotenv==1.0.0
fake-useragent==1.4.0
pyautogui==0.9.53

# Resume Processing
python-docx==0.8.11
pdfplumber==0.10.3
pypdf==3.17.1
reportlab==4.0.7

# Data & Analytics
pandas==2.1.3
plotly==5.18.0
matplotlib==3.8.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Monitoring & Logging
python-json-logger==2.0.7

# Development
black==23.12.0
ruff==0.1.8
mypy==1.7.1
